requirements.txt
flask>=2.3.0
joblib>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
tensorflow>=2.10.0
python-dotenv>=1.0.0
langdetect>=1.0.9
